+++ R/assess_SP.R
@@ -179,8 +179,8 @@ SP_Fox <- function(x = 1, Data, ...) {
   SP_args <- c(x = x, Data = Data, list(...))
   SP_args$start$n <- 1
   SP_args$fix_n <- TRUE
-
-  do.call(SP, SP_args)
+  
+  do.call2(SP, SP_args)
 }
 class(SP_Fox) <- "Assess"